wip. try intercepting diagnostics (fail)
tried to set a middleware which intercepts the diagnostics generated by
the system. It kind of works, firing on .ts but not on .svelte and seems
to fire only after an update..
but unlike the `provideFoldingRanges` middleware, it CANNOT execute a
command like "vscode.executeFoldingRangeProvider" and filter as we would
like. The vscode team explicitly noted that it:

> - Does not work with diagnostics errors. The VS Code API does not
support diagnostic providers that can 'pull' (request) diagnostics.
> -- https://code.visualstudio.com/api/language-extensions/embedded-languages#conclusion

so we cannot use it.

Next idea are:
- we are not intersted in diagnostics itself, let's try to tackle the
  'virtual document' approach.
- perplexity advise on using `.onDidChangeDiagnostics` to intercept the
  diagnostics.
  https://www.perplexity.ai/search/in-vscode-extensions-i-want-to-7fDTnZe1T1a1RC9QLE5N8g#1